add go

+++ tools/code_gen/java/output/TProgressBar.java
@@ -110,7 +110,7 @@ public class TProgressBar extends TWidget {
    *
    * @return 返回RET_OK表示成功，否则表示失败。
    */
- public  TRet setMax(int max)  {
+ public  TRet setMax(double max)  {
    return TRet.from(progress_bar_set_max(this != null ? (this.nativeObj) : 0, max));
  }
 
@@ -223,7 +223,7 @@ public class TProgressBar extends TWidget {
 static private native long progress_bar_create(long parent, int x, int y, int w, int h);
 static private native long progress_bar_cast(long widget);
 static private native int progress_bar_set_value(long widget, double value);
-static private native int progress_bar_set_max(long widget, int max);
+static private native int progress_bar_set_max(long widget, double max);
 static private native int progress_bar_set_format(long widget, String format);
 static private native int progress_bar_set_vertical(long widget, boolean vertical);
 static private native int progress_bar_set_show_text(long widget, boolean show_text);
